Separated the export authActions to a different file
